third_party/blink/web_tests/accessibility/listbox-enabled-states-expected.txt

Option 1
Option 2
Option 3
Option 4
Option 5

This tests that we report the correct enabled state on a listbox.

PASS axSelectElement.restriction is "none"
WARN: shouldBeEqualToString() expects string arguments
PASS axSelectElement.childAtIndex(0).isIgnored is true
PASS axSelectElement.childAtIndex(0).childAtIndex(0).restriction is "none"
PASS axSelectElement.childAtIndex(0).childAtIndex(1).restriction is "none"
PASS axSelectElement.childAtIndex(0).childAtIndex(2).restriction is "disabled"
PASS axSelectElement.childAtIndex(0).childAtIndex(3).restriction is "disabled"
PASS axSelectElement.childAtIndex(0).childAtIndex(4).restriction is "none"
PASS successfullyParsed is true

TEST COMPLETE
